Separate utilities related to pages
