+++ test/Test/Database/LSMTree/Internal/RunAcc.hs
@@ -16,6 +16,7 @@ import qualified Data.Vector.Primitive as VP
 import           Database.LSMTree.Internal.BlobRef (BlobSpan (..))
 import           Database.LSMTree.Internal.Entry
 import qualified Database.LSMTree.Internal.IndexCompact as Index
+import           Database.LSMTree.Internal.Page (PageNo (PageNo), singlePage)
 import qualified Database.LSMTree.Internal.PageAcc as PageAcc
 import qualified Database.LSMTree.Internal.PageAcc1 as PageAcc
 import qualified Database.LSMTree.Internal.RawBytes as RB
@@ -63,7 +64,7 @@ test_singleKeyRun =  do
     Just (fst (PageAcc.singletonPage k e)) @=? mp
     isJust mc @? "expected a chunk"
     True @=? Bloom.elem k b
-    Index.singlePage (Index.PageNo 0) @=? Index.search k ic
+    singlePage (PageNo 0) @=? Index.search k ic
 
 {-------------------------------------------------------------------------------
   PageAcc
